Practice 1/practice1/requirements.txt
fastapi
uvicorn
requests
numpy
opencv-python-headless
PyWavelets
Pillow
pytest